+++ tests/scriptlets/trusted-replace-node-text.test.js
@@ -72,6 +72,35 @@ test('simple case', (assert) => {
     }, 1);
 });
 
+test('simple case - check if quotes are correctly escaped', (assert) => {
+    const done = assert.async();
+
+    const nodeName = 'div';
+    const textMatch = 'alert';
+    const pattern = '/alert\\(\\\'test\\\'\\)/';
+    const replacement = 'alert(\\\'replaced\\\')';
+    const text = "alert('test')";
+    const expectedText = "alert('replaced')";
+
+    const nodeBefore = addNode('div', text);
+    const safeNodeBefore = addNode('a', text);
+
+    runScriptlet(name, [nodeName, textMatch, pattern, replacement]);
+
+    const nodeAfter = addNode('div', text);
+    const safeNodeAfter = addNode('span', text);
+    setTimeout(() => {
+        assert.strictEqual(nodeAfter.textContent, expectedText, 'text content should be modified');
+        assert.strictEqual(nodeBefore.textContent, expectedText, 'text content should be modified');
+
+        assert.strictEqual(safeNodeAfter.textContent, text, 'non-matched node should not be affected');
+        assert.strictEqual(safeNodeBefore.textContent, text, 'non-matched node should not be affected');
+
+        assert.strictEqual(window.hit, 'FIRED', 'hit function should fire');
+        done();
+    }, 1);
+});
+
 test('using matchers as regexes', (assert) => {
     const done = assert.async();